+++ macosx/tkMacOSXPrivate.h
@@ -321,6 +321,7 @@ VISIBILITY_HIDDEN
 @property int poolLock;
 @property int macOSVersion;
 @property Bool tkLiveResizeEnded;
+@property Bool tkWillExit;
 
 /*
  * Persistent state variables used by processMouseEvent.
@@ -402,12 +403,8 @@ VISIBILITY_HIDDEN
 {
 @private
     NSString *privateWorkingText;
-    Bool _tkNeedsDisplay;
-    NSRect _tkDirtyRect;
     NSTrackingArea *trackingArea;
 }
-@property Bool tkNeedsDisplay;
-@property NSRect tkDirtyRect;
 @property CGContextRef tkLayerBitmapContext;
 @end
 
@@ -417,8 +414,6 @@ VISIBILITY_HIDDEN
 @end
 
 @interface TKContentView(TKWindowEvent)
-- (void) addTkDirtyRect: (NSRect) rect;
-- (void) clearTkDirtyRect;
 - (void) generateExposeEvents: (NSRect) rect;
 - (void) tkToolbarButton: (id) sender;
 - (void) resetTkLayerBitmapContext;